required changes + clean code
