+++ packages/builder/package.json
@@ -30,13 +30,14 @@
     "postbuild": "node dist/schema-copy.js"
   },
   "dependencies": {
-    "@dotenv-run/webpack": "^1.0.1"
+    "@dotenv-run/webpack": "^1.2.1",
+    "@dotenv-run/esbuild": "^1.2.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.0.0",
     "@angular-devkit/schematics": "^17.0.0",
     "@angular-devkit/schematics-cli": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
+    "@angular/compiler": "^17.0.2",
     "cpy": "^8.1.2",
     "ts-node": "^10.8.0",
     "typescript": "~5.2.2"
