+++ packages/builder/src/builders/dev-server/index.ts
@@ -5,15 +5,15 @@ import {
 } from "@angular-devkit/architect";
 import {
   DevServerBuilderOptions,
-  executeDevServerBuilder,
   DevServerBuilderOutput,
+  executeDevServerBuilder,
 } from "@angular-devkit/build-angular";
-import { Observable, combineLatest } from "rxjs";
-import { switchMap } from "rxjs/operators";
-import { plugin } from "../plugin";
-import { plugin as esbuildPlugin } from "../esbuild-plugin";
-import { NgxEnvSchema } from "../ngx-env/ngx-env-schema";
+import { dotenvRun } from "@dotenv-run/esbuild";
+import { Observable, combineLatest, switchMap } from "rxjs";
+import { devServerIndexHtml } from "../../utils/esbuild-index-html";
 import { getProjectCwd } from "../../utils/project";
+import { NgxEnvSchema } from "../ngx-env/ngx-env-schema";
+import { plugin } from "../plugin";
 
 export const buildWithPlugin = (
   options: DevServerBuilderOptions & NgxEnvSchema,
@@ -40,9 +40,22 @@ export const buildWithPlugin = (
       };
       if (builderName === "@ngx-env/builder:application") {
         options.forceEsbuild = true;
-        return executeDevServerBuilder(options, context, null, {
-          buildPlugins: esbuildPlugin(ngxEnvOptions),
-        });
+        return executeDevServerBuilder(
+          options,
+          context,
+          {
+            indexHtml: async (content) =>
+              devServerIndexHtml(content, {
+                ...ngxEnvOptions,
+                appEnv: "NG_APP_ENV",
+              }),
+          },
+          {
+            buildPlugins: [
+              dotenvRun({ ...ngxEnvOptions, appEnv: "NG_APP_ENV" }),
+            ],
+          }
+        );
       } else {
         return executeDevServerBuilder(options, context, plugin(ngxEnvOptions));
       }
