+++ packages/builder/src/builders/plugin.ts
@@ -1,10 +1,7 @@
-import { DotenvRunOptions, Dict } from "@dotenv-run/core";
+import { DotenvRunOptions } from "@dotenv-run/core";
 import { DotenvRunPlugin } from "@dotenv-run/webpack";
 import type { Configuration } from "webpack";
-
-function escapeStringRegexp(str: string) {
-  return str.replace(/[|\\{}()[\]^$+*?.]/g, "\\$&").replace(/-/g, "\\x2d");
-}
+import { variablesReducer } from "../utils/variables-reducer";
 
 export function plugin(options: DotenvRunOptions, ssr = false) {
   const _options = { ...options, appEnv: "NG_APP_ENV" };
@@ -15,15 +12,6 @@ export function plugin(options: DotenvRunOptions, ssr = false) {
       webpackConfig.plugins.push(dotEnvPlugin);
       return webpackConfig;
     },
-    indexHtml: async (content: string) => {
-      return Object.keys(raw).reduce(
-        (html, key) =>
-          html.replace(
-            new RegExp("%" + escapeStringRegexp(key) + "%", "g"),
-            raw[key]
-          ),
-        content
-      );
-    },
+    indexHtml: async (content: string) => variablesReducer(content, raw),
   };
 }
