+++ packages/builder/src/utils/escape-string-regexp.ts
@@ -1,0 +1,3 @@
+export function escapeStringRegexp(str: string) {
+  return str.replace(/[|\\{}()[\]^$+*?.]/g, "\\$&").replace(/-/g, "\\x2d");
+}
